Refactor the way we have sctructure views
