+++ assets/js/components/Scoreboard.tsx
@@ -16,6 +16,7 @@ interface ScoreboardProps {
   game_data: string;
   recent_events_data: string;
   feature_flags_data?: string;
+  view?: string;
   pushEvent: (event: string, payload: any) => void;
   pushEventTo: (event: string, payload: any, selector: string) => void;
   handleEvent: (event: string, callback: (payload: any) => void) => void;
@@ -27,15 +28,20 @@ function Scoreboard({
   game_data,
   recent_events_data,
   pushEvent,
+  view,
 }: ScoreboardProps) {
   const object = JSON.parse(game_data);
   const recent_events_json = JSON.parse(recent_events_data);
   const game_state = (object.result as GameState) || DEFAULT_GAME_STATE;
   const recent_events: EventLog[] = recent_events_json.result || [];
   const sportId = game_state.sport_id ? game_state.sport_id : 'default';
-  const selectedView = game_state.view_settings_state
-    ? game_state.view_settings_state.view
-    : 'default';
+  const parsedView = JSON.parse(view || 'null');
+  const selectedView =
+    parsedView && parsedView !== 'default'
+      ? parsedView
+      : game_state.view_settings_state
+      ? game_state.view_settings_state.view
+      : 'default';
   const Component = ScoreboardRegistry[sportId];
   const isLoading = object.loading || false;
